+++ test/plugins/widget.js
@@ -57,7 +57,7 @@ describe('PluginWidget', function () {
       const status = widget.el.querySelector('.Placeholder-status')
       expect(widget).to.exist
       expect(widget.type).to.equal('placeholder')
-      expect(widget.el.firstChild.className).to.equal('Placeholder')
+      expect(widget.el.firstChild.classList.contains('Placeholder')).to.be.true
       expect(status.textContent).to.equal('Status')
     })
 
@@ -77,7 +77,7 @@ describe('PluginWidget', function () {
       ])
     })
 
-    it('updates widget props via updatePlaceholder', function (done) {
+    it('updates placeholder widget status via updatePlaceholder', function (done) {
       ed.on('media.update', function () {
         const widget = PluginWidget.widgets['0000']
         const status = widget.el.querySelector('.Placeholder-status')
@@ -85,7 +85,27 @@ describe('PluginWidget', function () {
         expect(status.textContent).to.equal('Status changed')
         done()
       })
-      ed.updatePlaceholder('0000', 'Status changed')
+      ed.updatePlaceholder('0000', {status: 'Status changed'})
+    })
+
+    it('updates placeholder widget errored true via updatePlaceholder', function (done) {
+      ed.on('media.update', function () {
+        const widget = PluginWidget.widgets['0000']
+        const el = widget.el.querySelector('.Placeholder')
+        expect(el.classList.contains('Placeholder-error')).to.be.true
+        done()
+      })
+      ed.updatePlaceholder('0000', {errored: true})
+    })
+
+    it('updates placeholder widget errored false via updatePlaceholder', function (done) {
+      ed.on('media.update', function () {
+        const widget = PluginWidget.widgets['0000']
+        const el = widget.el.querySelector('.Placeholder')
+        expect(el.classList.contains('Placeholder-error')).to.be.false
+        done()
+      })
+      ed.updatePlaceholder('0000', {errored: false})
     })
 
     it('changes widget type via setContent', function (done) {